test(sys): add sml test case
